[#2422] Move organization Zod schemas to src/schemas/models
Extract Zod schemas from src/types/organization.ts to their
canonical location at src/schemas/models/organization.ts.
Re-export schemas from types file for backward compatibility.

- Create src/schemas/models/organization.ts with all schemas
- Update src/types/organization.ts to keep only types/interfaces
- Export organization schemas from src/schemas/models/index.ts
- Update API endpoint to import from canonical location
